AC : long long型じゃないとダメだった。各変数が10^5のため

+++ atcoder207_pd/B-Hydrate.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main() {
-  int a, b, c, d;
+  long long a, b, c, d;
   int count = 0;
   cin >> a >> b >> c >> d;
